requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
spacy>=3.7.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
aiohttp>=3.9.3
click>=8.1.7
rich>=13.7.0
redis>=5.0.1
prometheus-client>=0.20.0
